Project Light finished
This project contains all the basic details to build an app using a button that will change the color of the background and also the text of the button when user click on it. It's a quickly practice to know how works the viewContent on an App and how to interact with the dynamic elements.